Refactoring src and test
